hbnu-swuniv-capstone-project-come-capstone25-CAPSTONE-TEMPLATE-main/003 Code/Agnet10/requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
openai>=1.0.0
python-dotenv>=1.0.0
requests>=2.31.0
pydantic>=2.7.0
